Split code in index.js file into separate files
